share challenge preview card
